Move grpc sevice into separate module. Fix a number of warnings
